pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sota-finder"
version = "0.1.0"
description = "Find and review candidate SOTA papers for biomedical ML datasets."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40",
    "google-generativeai>=0.8",
    "pdfplumber>=0.10",
    "pydantic>=2.0",
    "pymupdf>=1.24",
    "pypdf>=4.0",
    "requests>=2.28",
    "streamlit>=1.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"
